// modules/Livro/types.ts
export type LivroData = {
  codl?: number;
  titulo: string;
  editora: string;
  edicao: number;
  anopublicacao: string;
  valor: number;
};

// $table->string('titulo', 40)->index();
// $table->string('editora', 40);
// $table->smallInteger('edicao');
// $table->string('anopublicacao', 4);
// $table->float('valor', 8, 2);